refactor: move string related functions
